+++ app/src/main/res/layout/activity_main.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <woowacourse.paint.DrawingView
+        <woowacourse.paint.coustom.DrawingView
             android:id="@+id/drawing_view"
             android:layout_width="match_parent"
             android:layout_height="0dp"
